Improve parameter handling in cmdlets (#30)
* Improve parameter handling
* Check if the client exists before an update or delete operation

Closes #29

+++ src/Eryph.IdentityClient.Commands/RemoveEryphClientCommand.cs
@@ -1,29 +1,53 @@
-﻿using System.Management.Automation;
-using Eryph.IdentityClient.Models;
+﻿using Eryph.IdentityClient.Models;
 using JetBrains.Annotations;
+using System;
+using System.Management.Automation;
 
-namespace Eryph.IdentityClient.Commands
+namespace Eryph.IdentityClient.Commands;
+
+[PublicAPI]
+[Cmdlet(VerbsCommon.Remove,"EryphClient")]
+[OutputType(typeof(Client))]
+public class RemoveEryphClientCommand : IdentityCmdLet
 {
-    [PublicAPI]
-    [Cmdlet(VerbsCommon.Remove,"EryphClient")]
-    [OutputType(typeof(Client))]
-    public class RemoveEryphClientCommand : IdentityCmdLet
+    [Parameter(
+        Position = 0,
+        Mandatory = true,
+        ValueFromPipeline = true,
+        ValueFromPipelineByPropertyName = true)]
+    public string[] Id { get; set; }
+
+    /// <summary>
+    /// This parameter overrides the ShouldContinue call to force
+    /// the deletion of the eryph client. This parameter should always
+    /// be used with caution.
+    /// </summary>
+    [Parameter]
+    public SwitchParameter Force { get; set; }
+
+    private bool _yesToAll;
+    private bool _noToAll;
+
+    protected override void ProcessRecord()
     {
-        [Parameter(
-            Position = 0,
-            ValueFromPipeline = true,
-            ValueFromPipelineByPropertyName = true)]
-        public string[] Id { get; set; }
-
-        protected override void ProcessRecord()
+        var identityClient = Factory.CreateClientsClient();
+        foreach (var id in Id)
         {
-            var identityClient = Factory.CreateClientsClient();
-            foreach (var id in Id)
+            Client client;
+            try
             {
-                identityClient.Delete(id);
-
+                client = identityClient.Get(id);
             }
-        }
+            catch (Exception ex)
+            {
+                WriteError(new ErrorRecord(ex, "EryphClientNotFound", ErrorCategory.ObjectNotFound, id));
+                continue;
+            }
 
+            if (!Force && !ShouldContinue($"Eryph client '{client.Name}' (ID: {id}) will be deleted!", "Warning!", ref _yesToAll, ref _noToAll))
+                continue;
+
+            identityClient.Delete(id);
+        }
     }
 }
